Fix. Fixed issue that was causing the app server not to start.

+++ entrypoint.sh
@@ -1,5 +1,3 @@
 #!/bin/sh
-if [ ! -f /app/.env ]; then
-    cp /app/.env.example /app/.env
-fi
+if [ ! -f /app/.env ]; then cp /app/.env.example /app/.env; fi
 exec /app/server "$@"
